+++ CHANGELOG.md
@@ -89,6 +89,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Face3D: face_inside - now considers inners_contours
 - BoundingBox: point_belongs now considers bounds.
 - ContourMixin: delete_shared_contour_section
+- PlaneFace3D: merge_faces
 - Contour2D: divide
 
 ### Refactor
@@ -137,6 +138,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ClosedShel3D: intersection
 - Arcellipse2D: complementary
 - Contour2D: contours_from_edges.
+- PlaneFace3D: merge_faces
 - Contour2D: divide.
 v0.10.0 [Released 20/04/2023]